requirements.txt
Flask
flask-cors
opencv-python
insightface
onnxruntime-gpu
numpy